+++ frontend/src/app/(auth)/sign-up/page.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
+import SignUpForm from '@/components/features/auth/sign-up/SignUpForm'
 
 export default function Registration() {
-  return (
-    <div>page</div>
-  )
+  return <SignUpForm />
 }
