+++ frontend/src/components/features/auth/sign-up/SignUpForm.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import Footer from '../../footer/Footer'
+
+export default function SignUpForm() {
+  return (
+    <div className='flex flex-col  min-h-screen'>
+      <section className='flex-grow bg-gray-100 flex items-center justify-center p-4 md:p-14'>
+        Sign up form
+      </section>
+      <Footer />
+    </div>
+  )
+}
